--- Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/model/AnalysisReportDialog.kt
@@ -1,7 +1,0 @@
-package com.team.prezel.feature.report.impl.model
-
-enum class AnalysisReportDialog {
-    RE_RECORDING,
-    RE_WRITE_SCRIPT,
-    DELETE_REPORT,
-}
