--- Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/model/AnalysisReportUiMessage.kt
@@ -1,6 +1,0 @@
-package com.team.prezel.feature.report.impl.model
-
-internal enum class AnalysisReportUiMessage {
-    FETCH_REPORT_FAILED,
-    DELETE_REPORT_FAILED,
-}
