--- Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/model/PresentationInfoUiModel.kt
@@ -1,18 +1,0 @@
-package com.team.prezel.feature.report.impl.model
-
-import androidx.compose.runtime.Immutable
-import com.team.prezel.core.model.presentation.Audience
-import com.team.prezel.core.model.presentation.Category
-import com.team.prezel.core.model.presentation.Purpose
-import com.team.prezel.core.model.presentation.Style
-
-@Immutable
-internal data class PresentationInfoUiModel(
-    val category: Category,
-    val title: String,
-    val purpose: Purpose,
-    val style: Style,
-    val audience: Audience,
-    val analyzedAt: String,
-    val durationSeconds: Int,
-)
