--- Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/model/QuestionUiModel.kt
@@ -1,6 +1,0 @@
-package com.team.prezel.feature.report.impl.model
-
-data class QuestionUiModel(
-    val question: String,
-    val answer: String,
-)
