--- Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/model/ScriptAnalysisGraphData.kt
@@ -1,7 +1,0 @@
-package com.team.prezel.feature.report.impl.model
-
-internal data class ScriptAnalysisGraphData(
-    val spellingCount: Int,
-    val grammarCount: Int,
-    val totalErrorCount: Int,
-)
