--- Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/model/SpeedGraphData.kt
@@ -1,8 +1,0 @@
-package com.team.prezel.feature.report.impl.model
-
-import com.team.prezel.core.model.practice.RecordingSpeed
-
-internal data class SpeedGraphData(
-    val spm: Int,
-    val result: RecordingSpeed,
-)
