+++ Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/report/contract/AnalysisReportUiIntent.kt
@@ -1,0 +1,21 @@
+package com.team.prezel.feature.report.impl.report.contract
+
+import com.team.prezel.core.ui.base.UiIntent
+
+internal sealed interface AnalysisReportUiIntent : UiIntent {
+    data object ClickDelete : AnalysisReportUiIntent
+
+    data class ClickGrowthGraphItem(
+        val index: Int,
+    ) : AnalysisReportUiIntent
+
+    data object DismissDialog : AnalysisReportUiIntent
+
+    data object ClickDialogConform : AnalysisReportUiIntent
+
+    data object ClickReWriteScript : AnalysisReportUiIntent
+
+    data object ClickReRecording : AnalysisReportUiIntent
+
+    data object ClickFeedbackWrite : AnalysisReportUiIntent
+}
