+++ Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/report/model/AnalysisReportDialog.kt
@@ -1,0 +1,7 @@
+package com.team.prezel.feature.report.impl.report.model
+
+enum class AnalysisReportDialog {
+    RE_RECORDING,
+    RE_WRITE_SCRIPT,
+    DELETE_REPORT,
+}
