+++ Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/report/model/AnalysisReportUiMessage.kt
@@ -1,0 +1,6 @@
+package com.team.prezel.feature.report.impl.report.model
+
+internal enum class AnalysisReportUiMessage {
+    FETCH_REPORT_FAILED,
+    DELETE_REPORT_FAILED,
+}
