+++ Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/report/model/QuestionUiModel.kt
@@ -1,0 +1,6 @@
+package com.team.prezel.feature.report.impl.report.model
+
+data class QuestionUiModel(
+    val question: String,
+    val answer: String,
+)
