+++ Prezel/feature/report/impl/src/main/java/com/team/prezel/feature/report/impl/report/model/ScriptAnalysisGraphData.kt
@@ -1,0 +1,7 @@
+package com.team.prezel.feature.report.impl.report.model
+
+internal data class ScriptAnalysisGraphData(
+    val spellingCount: Int,
+    val grammarCount: Int,
+    val totalErrorCount: Int,
+)
